Attempts to add python support via multi buildpack
Using the multi buildpack requires a `.buildpacks` file which states
(in-order) the list of buildpacks to use. Because bosh (as part of the
release) uses unix globbing to grab the list of files to upload as part
of your release (see your packages/spec file), it by default ignores
hidden files. Because of this, we have to call it `buildpacks` instead
and as part of the bosh deploy, copy this to the correct name.

We also could not use the default multi buildpack because the
cloudfoundry python buildpack uses submodules, and the buildpack does
not initialize submodules. We forked the multi buildpack to perform
this, and are using the multi buildpack at the relevant fork's branch.

The python buildpack installs python into the `/app/.heroku/python`
directory, including header files in
`/app/.heroku/python/include/python2.7` and `/app/.heroku/python/lib`.
This non-standard location then needs to be added to the gcc flags when
compiling the https://github.com/s-u/rpython/blob/master/src/py.c file
in `rpython2`. To do this, we append onto the `PKG_CPPFLAGS` and
`PKG_LIBS` and `LD_LIBRARY_PATH` environment variables before
installing `rpython2`.

For some reason, the above process failed if we overrode those
environment variables before installing other R dependencies, so we put
it at the end of the script, immediately before the `rpython2`
installation.

`rpython2` requires that the python be installed using `--enable-shared`
in the configure step. The heroku python buildpack python-2.7.8 (which is
the  default in the python buildpack), and the python-2.7.8-shared does
not use this flag, so we had to downgrade the python installation (in
`runtime.txt`) to python-2.7.6 which we know is configured properly.

[#81407056]